requirements.txt
python-telegram-bot[webhooks]>=21.5
openai==0.28.1